[FIX] web_editor: revert copy/paste feature of the editor
The feature was introduced and fixed by following commits:
- https://github.com/odoo/odoo/commit/2aa997634f94aba21ddf05dfa7533f6ce5b10afe
- https://github.com/odoo/odoo/commit/12fa8954176682b92822aa2121f886ea51a28e0d
- https://github.com/odoo/odoo/commit/e31ee7a4751f442057a9abe7f710ff4f74dcba9c

The feature is however still buggy in lots of cases as it is a bad
heuristic that cannot handle simple pasting cases. The feature was
initially introduced to allow pasting PDF content without having to
remove all the unwanted line breaks this induces. As this feature is
not that important and introduced more bugs than it solved, the feature
is deleted by this commit. If we want that feature again, maybe it is
worth considering an "Import from PDF" button or something like that.

See https://github.com/odoo/odoo/issues/15665