[FIX] crm: A closed lead doesn't mean there is nothing more to do
Before this commit, because of 8d1217d and 23418f9 and 906994cacd40bfb70d7aaf4ec6f12f6b0b32ca8a
closed crm.lead with late activities wouldn't show.

Aware that this is essentially a choice of non-crucial behavior, there are some precautionary arguments.
- The three commits above make the choice of having a "clean" backlog:
late activities are not polluted by crm leads that are already won
- The choice of this commit is to count activities as late regardless of the state of the lead:
The use case: Close a deal with a partner > lead is won and closed
But it doesn't mean you don't have to call the partner back in a few days to actually sign a paper, organize some sort of meeting

The argument: it is better to not miss deadlines eventhough they might accumulate, rather than forgetting about them altogether
(but you don't script something to put all your emails as read when they arrive, do you ?)

OPW 744940