[FIX][IMP] sale: set sale_warn as non required
When reinitializing modules it is impossible to create partners if
sale is installed because sale_warn field has a required column
but no default value if the module does not depend from sale. Use
case is trying to run test_mail tests with sale installed.

As this field is not business critical it is now not required. As there
is a default value behavior should not change for users.

See also d2533fefeec78710d2c8700f1ab15a5a00205a53 and 6326773162caf37655adcfd72286ad9b776b0a34.