[FIX] doc: informal sub-params
Revert f06d224fb23b7e36b4f6ddb592d4c63c36e67dca which reverted
86b9c97b745bec22c8ffe31112a2c0152f6d9478.

Issue was the support for sub-parameters (aka @param {type} foo.bar)
where the "parent" parameter was not formal (only present in a
function's docstring, but not in the function's parameters list),
subtype creation would fail to find a root parameter to attach the
sub-parameter to.

This has not actually been fixed, but for the specific case where the
formal parameters list is *completely empty* we fall back on a list of
the docstring's parameters, which works well enough for this case.